+++ backend/src/main/java/org/vaccom/vcmgt/action/NguoiTiemChungAction.java
@@ -36,7 +36,7 @@ public interface NguoiTiemChungAction {
 	public List<NguoiTiemChung> searchNguoiTiemChung(String cmtcccd, Integer nhomdoituong, String ngaydangki,
 			String hovaten, Long diabancosoid, String cosoytema, Integer tinhtrangdangky, Integer kiemtratrung,
 			Integer page, Integer size, String tinhthanhma, String tinhthanhten, String quanhuyenma
-			,  String quanhuyenten, String phuongxama, String phuongxaten);
+			,  String quanhuyenten, String phuongxama, String phuongxaten, Boolean isDatTieuChuan);
 
 	public long countNguoiTiemChung(String cmtcccd, Integer nhomdoituong, String ngaydangki, String hovaten,
 			Long diabancosoid, String cosoytema, Integer tinhtrangdangky, Integer kiemtratrung, String tinhthanhma, String tinhthanhten, String quanhuyenma
@@ -61,4 +61,5 @@ public interface NguoiTiemChungAction {
 	public NguoiTiemChung update(NguoiTiemChung nguoiTiemChung);
 
 	public NguoiTiemChung findBycongDanID(long id);
+
 }