+++ backend/src/main/java/org/vaccom/vcmgt/service/NguoiTiemChungService.java
@@ -21,6 +21,7 @@ public interface NguoiTiemChungService {
 
 	public long countAll();
 
+
 	public NguoiTiemChung findById(long id);
 
 	public NguoiTiemChung findByMaQR(String MaQR);
@@ -41,7 +42,7 @@ public interface NguoiTiemChungService {
 
 	public List<NguoiTiemChung> searchNguoiTiemChung(String cmtcccd, Integer nhomdoituong, String ngaydangki,
 			String hovaten, Long diabancosoid, String cosoytema, Integer tinhtrangdangki, Integer kiemtratrung, Integer page, Integer size, String tinhthanhma, String tinhthanhten, String quanhuyenma
-			,  String quanhuyenten, String phuongxama, String phuongxaten);
+			,  String quanhuyenten, String phuongxama, String phuongxaten, Boolean isDatTieuChuan);
 
 	public ResultSearchDto<NguoiTiemChung> search(NguoiTiemChungDto nguoiTiemChungDto, int page, int size);
 	public ResultSearchDto<CongDan> searchDaTiem(NguoiTiemChungDto nguoiTiemChungDto, int page, int size);