+++ backend/src/main/java/org/vaccom/vcmgt/service/PhieuHenTiemService.java
@@ -4,6 +4,9 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import org.vaccom.vcmgt.dto.NguoiTiemChungDto;
+import org.vaccom.vcmgt.dto.PhieuHenTiemDto;
+import org.vaccom.vcmgt.dto.ResultSearchDto;
 import org.vaccom.vcmgt.entity.PhieuHenTiem;
 
 /**